# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order POS API",
        "description": "API para la gestión de pedidos, productos y usuarios en puntos de venta.",
        "contact": {
            "name": "Jhonatan Guerrero",
            "email": "jhonatanguerrero@outlook.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos habilitados",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Lista de productos habilitados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto X"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
...
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener producto habilitado por ID",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado o no habilitado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto habilitado",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo nombre"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 199.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado o deshabilitado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Deshabilitar (soft delete) producto",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto deshabilitado y soft deleted"
                    },
                    "404": {
                        "description": "Producto no encontrado o ya eliminado"
                    }
                }
            }
        },
        "/api/products/trashed/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener producto eliminado (soft deleted) por ID",
                "operationId": "fd8f4d666f3b1d0af148a166941bcab6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto eliminado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deleted product not found"
                    }
                }
            }
        },
        "/api/products/{id}/restore": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Restaurar producto eliminado (soft deleted)",
                "operationId": "cac6708948b75b3fcc612aed6626824c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto eliminado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto restaurado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product restored successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deleted product not found"
                    }
                }
            }
        },
        "/api/products/{id}/force": {
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto permanentemente",
                "operationId": "3c95bd58069e5b3d7356552bc209e53a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto eliminado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado permanentemente"
                    },
                    "404": {
                        "description": "Deleted product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Producto",
                "description": "Modelo que representa un producto del sistema",
                "required": [
                    "id",
                    "name",
                    "price",
                    "enabled"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del producto",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Café Americano"
                    },
                    "price": {
                        "description": "Precio del producto",
                        "type": "number",
                        "format": "float",
                        "example": 15.5
                    },
                    "enabled": {
                        "description": "Indica si el producto está habilitado",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T12:34:56Z"
                    },
                    "deleted_at": {
                        "description": "Fecha de eliminación (null si activo)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}